refactor: 파일 위치 및 변수명 변경 등
