+++ backend/src/main/java/com/multi/mariage/product/vo/FoodRateVO.java
@@ -1,6 +1,5 @@
-package com.multi.mariage.product.vo.filter;
+package com.multi.mariage.product.vo;
 
-import com.multi.mariage.category.domain.FoodCategory;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Data;
@@ -9,20 +8,20 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Data
 public class FoodRateVO {
-    private int id;
+    private int foodId;
     private String category;
     private double avgFoodRate;
 
     @Builder
-    private FoodRateVO(int id, String category, double avgFoodRate) {
-        this.id = id;
+    private FoodRateVO(int foodId, String category, double avgFoodRate) {
+        this.foodId = foodId;
         this.category = category;
         this.avgFoodRate = avgFoodRate;
     }
 
-    public static FoodRateVO from(int id, String category, double avgFoodRate) {
+    public static FoodRateVO from(int foodId, String category, double avgFoodRate) {
         return FoodRateVO.builder()
-                .id(id)
+                .foodId(foodId)
                 .category(category)
                 .avgFoodRate(avgFoodRate)
                 .build();
