+++ backend/src/main/java/com/multi/mariage/product/vo/ReviewRateVO.java
@@ -1,4 +1,4 @@
-package com.multi.mariage.review.vo;
+package com.multi.mariage.product.vo;
 
 import lombok.AccessLevel;
 import lombok.Builder;
